+++ module/iam/register_test.go
@@ -32,8 +32,8 @@ func userStatusAPI(userID string) string {
 }
 
 type ListResponse[T any] struct {
-	Items []T   `json:"items"`
-	Total int64 `json:"total"`
+	Items []T `json:"items"`
+	Total int `json:"total"`
 }
 
 func init() {
